+++ CompanionPackages~/com.playserv.schema-tool/Tools~/SchemaTool/Source/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 
-[assembly: AssemblyVersion("0.6.8.0")]
-[assembly: AssemblyFileVersion("0.6.8.0")]
-[assembly: AssemblyInformationalVersion("0.6.8")]
+[assembly: AssemblyVersion("0.6.9.0")]
+[assembly: AssemblyFileVersion("0.6.9.0")]
+[assembly: AssemblyInformationalVersion("0.6.9")]